Rename transformValue to transformer to reflect its multiple exports
